+++ frontend/src/libs/import-export/index.ts
@@ -5,5 +5,5 @@ export * from "./match-by-utorid-or-name";
 export * from "./normalize-import";
 export * from "./validate";
 export * from "./prepare-spreadsheet";
-export * from "./prepare-json";
+export * from "./prepare-minimal";
 export * from "./normalize-ddah-import";
