+++ frontend/src/libs/import-export/prepare-minimal.ts
